GasInSight.Hub/requirements.txt
azure-servicebus==7.11.4
python-dotenv==1.0.1
pydantic==2.6.1
faker==22.6.0
